[Truffle] Fix some mx warnings.

+++ truffle/src/main/java/org/jruby/truffle/util/Sprintf.java
@@ -216,7 +216,7 @@ public class Sprintf {
         byte[] bytes = new byte[len];
         if (upper) {
             for (int i = len; --i >= 0; ) {
-                int b = (byte)((int)s.charAt(i) & (int)0xff);
+                int b = (byte)((int)s.charAt(i) & 0xff);
                 if (b >= 'a' && b <= 'z') {
                     bytes[i] = (byte)(b & ~0x20);
                 } else {
@@ -225,7 +225,7 @@ public class Sprintf {
             }
         } else {
             for (int i = len; --i >= 0; ) {
-                bytes[i] = (byte)((int)s.charAt(i) & (int)0xff);
+                bytes[i] = (byte)((int)s.charAt(i) & 0xff);
             }
         }
         return bytes;
